Implement button content on html
I decided my current implementation was overbloated and overcomplicated
so this branch is where I will rewrite a good portion of the code.
Beginning with adding the button elements directly on the HTML pages
instead of in memory. This not only reduces the code needed in the
javascript file but also makes referencing and using buttons much easier
than I originally implemented.