Fix display overflow issue
After testing the keyboard support, it became apparent that any values
that reached a length of 9 would not display valid calculations applied.
For e.g. "555555555 + 5" should display 555555560 but instead the
display remained unchanged.
Had to rearrange the numberhandler function to have the function check
for afterOper boolean first before checking the length of the display.