Clean and simplify roundDisplay function
Made changes to remove redundant code and make it more efficient.
For example, removing the if else statement for positive number and
negative number casing. On realising the negative symbol in the display
would be considered into the string length, I had repeated a whole block
of code for no reason.
As well as this, efficient processing in using the slice method over the
substring method in some cases was implemented.